+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.0.0
+
 - Add `EctoTemp.Factory.uuid/0`.
 - Do not automatically append `_temp` to table names.